[risk=no] Factor out workspace stubs to avoid circular dep
