+++ ui/src/app/cohort-search/gender-chart/gender-chart.component.spec.tsx
@@ -2,7 +2,7 @@ import {shallow} from 'enzyme';
 import * as React from 'react';
 
 import {currentWorkspaceStore} from 'app/utils/navigation';
-import {workspaceDataStub} from 'testing/stubs/workspaces-api-stub';
+import {workspaceDataStub} from 'testing/stubs/workspaces';
 import {GenderChart} from './gender-chart.component';
