+++ ui/src/app/cohort-search/search-group-list/search-group-list.component.spec.tsx
@@ -5,7 +5,7 @@ import {registerApiClient} from 'app/services/swagger-fetch-clients';
 import {currentWorkspaceStore} from 'app/utils/navigation';
 import {CohortBuilderApi} from 'generated/fetch';
 import {CohortBuilderServiceStub} from 'testing/stubs/cohort-builder-service-stub';
-import {workspaceDataStub} from 'testing/stubs/workspaces-api-stub';
+import {workspaceDataStub} from 'testing/stubs/workspaces';
 import {SearchGroupList} from './search-group-list.component';
 
 describe('SearchGroupList', () => {
