+++ ui/src/app/components/help-sidebar.spec.tsx
@@ -11,7 +11,7 @@ import defaultServerConfig from 'testing/default-server-config';
 import {waitForFakeTimersAndUpdate, waitOneTickAndUpdate} from 'testing/react-test-helpers';
 import {CohortAnnotationDefinitionServiceStub} from 'testing/stubs/cohort-annotation-definition-service-stub';
 import {CohortReviewServiceStub, cohortReviewStubs} from 'testing/stubs/cohort-review-service-stub';
-import {workspaceDataStub} from 'testing/stubs/workspaces-api-stub';
+import {workspaceDataStub} from 'testing/stubs/workspaces';
 import colors from 'app/styles/colors';
 import {runtimeStore} from 'app/utils/stores';
 import {HelpSidebar} from './help-sidebar';
