+++ ui/src/app/pages/workspace/workspace-card.spec.tsx
@@ -5,7 +5,8 @@ import * as React from 'react';
 import {registerApiClient, workspacesApi} from 'app/services/swagger-fetch-clients';
 import {Profile, ProfileApi, WorkspacesApi, WorkspaceAccessLevel} from 'generated/fetch';
 import {ProfileApiStub, ProfileStubVariables} from 'testing/stubs/profile-api-stub';
-import {workspaceStubs, WorkspacesApiStub, userRolesStub} from 'testing/stubs/workspaces-api-stub';
+import {workspaceStubs, userRolesStub} from 'testing/stubs/workspaces';
+import {WorkspacesApiStub} from 'testing/stubs/workspaces-api-stub';
 import {serverConfigStore, userProfileStore} from 'app/utils/navigation';
 import {waitOneTickAndUpdate} from 'testing/react-test-helpers';
 import {WorkspaceCard} from './workspace-card';
