+++ ui/src/testing/stubs/cohorts-api-stub.ts
@@ -12,7 +12,7 @@ import {
 import {CohortListResponse} from 'generated/fetch/api';
 import {stubNotImplementedError} from 'testing/stubs/stub-utils';
 import {convertToResources} from './resources-stub';
-import {WorkspaceStubVariables} from './workspace-service-stub';
+import {WorkspaceStubVariables} from './workspaces';
 
 export let DEFAULT_COHORT_ID = 1;
 export let DEFAULT_COHORT_ID_2 = 2;
